Split bursts mono-module into separate files
