+++ src/bursts/getExtraAttackDamageFramesEntry.ts
@@ -1,41 +1,7 @@
-import {
-	IBraveBurst,
-	IBurstLevelEntry,
-	ProcEffect,
-	IBurstDamageFramesEntry,
-} from './datamine-types';
-import { getEffectId, isAttackingProcId } from './buffs';
-import { KNOWN_PROC_ID } from './constants';
-
-/**
- * @description Grab the level entry of a burst at a given level (or the last level if no level is given)
- * @param burst Burst to get level entry from
- * @param level Optional 1-indexed level to get; if not specified, the last level of the burst is used.
- * @returns the level entry of a burst at a given level (or last level if no level is given) if it exists, undefined otherwise
- */
-export function getLevelEntryForBurst (burst: IBraveBurst, level?: number): IBurstLevelEntry | undefined {
-	const burstEffectsByLevel = (burst && Array.isArray(burst.levels)) ? burst.levels : [];
-	let levelIndex: number;
-	if (level !== null && !isNaN(level as number)) {
-		// 1-indexed
-		levelIndex = (+(level as number) - 1);
-	} else {
-		// default to last entry in burst
-		levelIndex = burstEffectsByLevel.length - 1;
-	}
-	return burstEffectsByLevel[levelIndex];
-}
-
-/**
- * @description Grab the effects at the level entry of a burst at a given level (or the last level if no level is given)
- * @param burst Burst to get effects from
- * @param level Optional 1-indexed level to get entries from; if not specified, the last level of the burst is used.
- * @returns the effects at the level entry of a burst at a given level (or last level if no level is given) if it exists, an empty array otherwise
- */
-export function getEffectsForBurst (burst: IBraveBurst, level?: number): ProcEffect[] {
-	const levelEntry = getLevelEntryForBurst(burst, level);
-	return (levelEntry && Array.isArray(levelEntry.effects)) ? levelEntry.effects : [];
-}
+import { IBurstDamageFramesEntry } from '../datamine-types';
+import getEffectId from '../buffs/getEffectId';
+import { KNOWN_PROC_ID } from '../constants';
+import isAttackingProcId from '../buffs/isAttackingProcId';
 
 /**
  * @description Get the extra attack damage frames entry based on the damage frames of a burst. Also apply the given effect delay to the resulting damage frames entry.
@@ -43,7 +9,7 @@ export function getEffectsForBurst (burst: IBraveBurst, level?: number): ProcEff
  * @param effectDelay optional effect delay to apply to the resulting damage frames entry
  * @returns damage frames entry whose frames are based on the input damage frames
  */
-export function getExtraAttackDamageFramesEntry (
+export default function getExtraAttackDamageFramesEntry (
 	damageFrames: IBurstDamageFramesEntry[],
 	effectDelay = '0.0/0',
 ): IBurstDamageFramesEntry {
